session3

+++ api/database/models.py
@@ -1,25 +1,25 @@
 from database.database import base
-from sqlalchemy import (DATE, TEXT, TIME, TIMESTAMP, Boolean, Column,
+from sqlalchemy import (DATE, TEXT, TIME, TIMESTAMP, Boolean, Column, Float,
                         ForeignKey, Integer, String)
 
 
 class Roles(base):
     __tablename__ = "roles"
-    
+
     ID = Column(Integer, primary_key=True)
     Title = Column(TEXT)
 
 
 class CabinTypes(base):
     __tablename__ = "cabintypes"
-    
+
     ID = Column(Integer, primary_key=True)
     Name = Column(TEXT)
 
 
 class Countries(base):
     __tablename__ = "countries"
-    
+
     ID = Column(Integer, primary_key=True)
     Name = Column(TEXT)
 
@@ -56,23 +56,23 @@ class Users(base):
 
 class Logs(base):
     __tablename__ = "logs"
-    
+
     ID = Column(Integer, primary_key=True)
-    
+
     Error = Column(TEXT)
     UserID = Column(ForeignKey(Users.ID))
     Time = Column(TIMESTAMP)
-    
+
 
 class Tokens(base):
     __tablename__ = "tokens"
-    
+
     ID = Column(Integer, primary_key=True)
     Token = Column(TEXT)
-    
+
     CreateTime = Column(TIMESTAMP)
     DeletionTime = Column(TIMESTAMP)
-    
+
     Active = Column(Boolean)
     UserID = Column(ForeignKey(Users.ID))
 
@@ -124,7 +124,7 @@ class Schedules(base):
     RouteID = Column(ForeignKey(Routes.ID))
 
     FlightNumber = Column(TEXT)
-    EconomyPrice = Column(Integer)
+    EconomyPrice = Column(Float)
 
     Confirmed = Column(Integer)
